installed api of current weather call
