+++ src/Form.js
@@ -6,37 +6,37 @@ import DisplayHourlyWeather from './DisplayHourlyWeather';
 import Display10DaysWeather from './Display10DaysWeather';
 import axios from 'axios';
 
-const Form = () => {
+const Form = (props) => {
 
-   const [ cityChoice, setCityChoice ] = useState("");
-   const [ weatherResult, setWeatherResult ] = useState({});
+   // const [ cityChoice, setCityChoice ] = useState("");
+   // const [ weatherResult, setWeatherResult ] = useState({});
 
-   const handleCityChange = (event) => {
-      setCityChoice(event.target.value);
-   }
+   // const handleCityChange = (event) => {
+   //    setCityChoice(event.target.value);
+   // }
 
-   const handleSubmit = (event) => {
-      event.preventDefault();
+   // const handleSubmit = (event) => {
+   //    event.preventDefault();
       
-      axios({
-         url: `https://api.openweathermap.org/data/2.5/weather/`,
-         method: "GET",
-         dataResponse: "json",
-         params: {
-            q: cityChoice,
-            appid: "3d828a8d5ff245862af24fb2c5883de1",
-            units: "metric",
-         }
-      }).then((res) => {
-         setWeatherResult(res.data);
-      });
-      setCityChoice('');
-   }
+   //    axios({
+   //       url: `https://api.openweathermap.org/data/2.5/weather/`,
+   //       method: "GET",
+   //       dataResponse: "json",
+   //       params: {
+   //          q: cityChoice,
+   //          appid: "3d828a8d5ff245862af24fb2c5883de1",
+   //          units: "metric",
+   //       }
+   //    }).then((res) => {
+   //       setWeatherResult(res.data);
+   //    });
+   //    setCityChoice('');
+   // }
 
    return (
       <section>
          <h2>The forecast calls for...</h2>
-            <form onSubmit = {handleSubmit}>
+            <form action="submit" onSubmit = {event => props.handleSubmit(event, props.cityChoice)}>
                <label htmlFor="city">Search by city:</label>
                <input 
                   placeholder="Toronto"
@@ -44,10 +44,10 @@ const Form = () => {
                   type="text" 
                   id="city" 
                   name="city" 
-                  onChange = {handleCityChange} 
-                  value = {cityChoice} 
+                  onChange = {event => props.handleCityChange(event)} 
+                  value = {props.cityChoice} 
                />
-               <Link to="/"><button type="submit">Search</button></Link>
+               <button type="submit">Search</button>
             </form>
          <Link to="/hourly"><button>Hourly Forecast</button></Link>
          <Link to ="/10days"><button>10 Day Forecast</button></Link>
